Canvas selector wasn't working, now it is.
The syntax for selecting the canvas with the ID 'main' was wrong, there
shouldn't have been a space. I didn't see a reason to limit the style to
only the 'main' canvas, as the selector really doesn't do anything, so I
just made it affect all canvases present on the page. This might change
later though.