Fix bug where locks persist across gesture change
This occurred when a user either started by dragging an item then added
other touch inputs to perform another gesture, or when their first touch
input that landed on the screen when performing a multi-input gesture
happened to hit an item. The lock on the item would persist despite the
change in gestures.

Now, the lock is released when a change in gesture is detected. The
change is detected by the gesture library as any time the number of
active and moving inputs increases beyond 1.