Add custom.css to .gitignore
Adds some styling for the docs, but I keep needing to wipe the docs/
directory and regenerate when I muck around with the file structure and
the way the various classes, etc., are linked. Keeping this file at this
level saves a little bit of time, as I can just copy it back down into
the docs/styles/ directory, without having to save a copy first.