Fix bug: jumpy gestures after inputs cleared
When the inputs associated with a view are cleared, there was a bug
where the next update from the GestureController would be jumpy. This
was fixed by making sure that the convenience fields of the State object
were appropriately updated after the clear and before the 'cancel' phase
was evaluated.