Remove Publisher from code
Emit updates automatically.

This was a carefully considered decision. Basically I came to the
conclusion that having scheduled updates results in traffic surges 60
times a second, and can introduce delays and jitter. Better to send
updates immediately. The drawing procedure can still be limited to once
every 60 seconds though. That is a wholly separate operation.