Fix jitter when manipulating items or views
It was particularly bad while manipulating views. This was caused by the
emission of several updates when only one would do. This was solved by
the following:
- Make view groups mix Lockable and Transformable2D instead of
Interactable, essentially removing the publish() step when
transformations were applied, which was a multiplicative source of
repeats, hence the extreme jitter when manipulating views in server-side
gesture mode.
- Adjust the Publishable mixin such that when publish() is called, it
waits until the end of the current turn of the event loop and then emits
a single update. Previously, updates were fired immediately, one per
transformation. This is more efficient, and removes a lot of the
jitter!!