Switch to Touch events, add smoothing (srvr gstrs)
Clients now forward touch events instead of pointer events, thus
ensuring that a collected touch event gets processed in one go on the
server side. I also added smoothing to the Pinch and Rotate gestures.

The smoothing is slightly different from what is in westures, as it
simply averages the next two emits instead of trying to do something
fancier.