Fix bug preventing item locking in some cases
Specifically, if a user's first interaction consisted of more than one
touch, all of which started within the same 60th of a second. The
locking mechanism wouldn't kick in, and interaction would completely
fail. This only seemd to crop up with server-side gestures, presumably
due to the update batching system, and because user devices send a
'touchstart' for each individual touch, so the aggregation causing the
problem simply didn't occur for client-side gestures.

The fix was to check if the user had a locked item when starting
tracking, rather than checking the number of inputs. A more robust
solution might be to set a flag when the number of inputs drops to 0.